+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # Changelog
 
 ## **1.0.1** - *2019-xx-xx* (next)
-* `Dashboard/Gallery` - success toasts added, cards loading behavior added;
+* `Dashboard/Gallery`:
+    * success toasts added;
+    * cards loading behavior added;
+    * photosets appearance improved;
 * `Dashboard` - global loading overlay added;
 
 ## **1.0.0** - *2019-10-26*